+++ Source/Scene/MapboxStyleImageryProvider.js
@@ -19,7 +19,7 @@ var defaultCredit = new Credit(
  * @property {Resource|String} [url='https://api.mapbox.com/styles/v1/'] The Mapbox server url.
  * @property {String} [username='mapbox'] The username of the map account.
  * @property {String} styleId The Mapbox Style ID.
- * @property {String} [accessToken] The public access token for the imagery.
+ * @property {String} accessToken The public access token for the imagery.
  * @property {Number} [tilesize=512] The size of the image tiles.
  * @property {Boolean} [scaleFactor] Determines if tiles are rendered at a @2x scale factor.
  * @property {Ellipsoid} [ellipsoid] The ellipsoid.  If not specified, the WGS84 ellipsoid is used.
@@ -55,6 +55,13 @@ function MapboxStyleImageryProvider(options) {
   //>>includeStart('debug', pragmas.debug);
   if (!defined(styleId)) {
     throw new DeveloperError("options.styleId is required.");
+  }
+  //>>includeEnd('debug');
+
+  var accessToken = MapboxApi.getAccessToken(options.accessToken);
+  //>>includeStart('debug', pragmas.debug);
+  if (!defined(accessToken)) {
+    throw new DeveloperError("options.accessToken is required.");
   }
   //>>includeEnd('debug');
 
@@ -148,13 +155,8 @@ function MapboxStyleImageryProvider(options) {
     defaultValue(options.url, "https://api.mapbox.com/styles/v1/")
   );
 
-  var accessToken = MapboxApi.getAccessToken(options.accessToken);
   this._styleId = styleId;
   this._accessToken = accessToken;
-
-  this._accessTokenErrorCredit = Credit.clone(
-    MapboxApi.getErrorCredit(options.accessToken)
-  );
 
   var tilesize = defaultValue(options.tilesize, 512);
   this._tilesize = tilesize;
@@ -405,9 +407,7 @@ Object.defineProperties(MapboxStyleImageryProvider.prototype, {
  * @exception {DeveloperError} <code>getTileCredits</code> must not be called before the imagery provider is ready.
  */
 MapboxStyleImageryProvider.prototype.getTileCredits = function (x, y, level) {
-  if (defined(this._accessTokenErrorCredit)) {
-    return [this._accessTokenErrorCredit];
-  }
+  return undefined;
 };
 
 /**