+++ Specs/TypeScript/index.ts
@@ -108,8 +108,11 @@ imageryProvider = new OpenStreetMapImageryProvider({ url: "" });
 imageryProvider = new TileMapServiceImageryProvider({ url: "" });
 imageryProvider = new GridImageryProvider({ url: "" });
 imageryProvider = new IonImageryProvider({ assetId: 2 });
-imageryProvider = new MapboxImageryProvider({ mapId: "" });
-imageryProvider = new MapboxStyleImageryProvider({ styleId: "" });
+imageryProvider = new MapboxImageryProvider({ mapId: "", accessToken: "" });
+imageryProvider = new MapboxStyleImageryProvider({
+  styleId: "",
+  accessToken: "",
+});
 imageryProvider = new SingleTileImageryProvider({ url: "" });
 imageryProvider = new TileCoordinatesImageryProvider();
 imageryProvider = new UrlTemplateImageryProvider({ url: "" });